#7: refactor messages from QueueController to new MessageController
